LICSENSE
QdPieDrury Script License 1.0

Copyright (c) 2025 QdPieDrury

Permission is hereby granted to use, modify, and distribute this software freely,  
provided that the original author (QdPieDrury) is credited.

You may NOT repost, redistribute, or publish this script claiming it as your own work.

This software is provided "as is", without warranty of any kind. The author is not liable for any damages arising from its use.